+++ src/extension/internal/bitmap/swirl.h
@@ -17,7 +17,7 @@ namespace Bitmap {
 class Swirl : public ImageMagick
 {
 private:
-	float _degrees;
+	int _degrees;
 public:
 	void applyEffect(Magick::Image *image) override;
 	void refreshParameters(Inkscape::Extension::Effect *module) override;
